+++ crates/hyperswitch_domain_models/src/payouts/payouts.rs
@@ -90,6 +90,8 @@ pub struct Payouts {
     pub profile_id: String,
     pub status: storage_enums::PayoutStatus,
     pub confirm: Option<bool>,
+    pub payout_link_id: Option<String>,
+    pub client_secret: Option<String>,
     pub priority: Option<storage_enums::PayoutSendPriority>,
 }
 
@@ -116,6 +118,8 @@ pub struct PayoutsNew {
     pub profile_id: String,
     pub status: storage_enums::PayoutStatus,
     pub confirm: Option<bool>,
+    pub payout_link_id: Option<String>,
+    pub client_secret: Option<String>,
     pub priority: Option<storage_enums::PayoutSendPriority>,
 }
 
@@ -145,6 +149,8 @@ impl Default for PayoutsNew {
             profile_id: String::default(),
             status: storage_enums::PayoutStatus::default(),
             confirm: None,
+            payout_link_id: Option::default(),
+            client_secret: Option::default(),
             priority: None,
         }
     }
@@ -165,6 +171,7 @@ pub enum PayoutsUpdate {
         profile_id: Option<String>,
         status: Option<storage_enums::PayoutStatus>,
         confirm: Option<bool>,
+        payout_type: Option<storage_enums::PayoutType>,
     },
     PayoutMethodIdUpdate {
         payout_method_id: String,
@@ -196,6 +203,7 @@ pub struct PayoutsUpdateInternal {
     pub status: Option<storage_enums::PayoutStatus>,
     pub attempt_count: Option<i16>,
     pub confirm: Option<bool>,
+    pub payout_type: Option<common_enums::PayoutType>,
 }
 
 impl From<PayoutsUpdate> for PayoutsUpdateInternal {
@@ -214,6 +222,7 @@ impl From<PayoutsUpdate> for PayoutsUpdateInternal {
                 profile_id,
                 status,
                 confirm,
+                payout_type,
             } => Self {
                 amount: Some(amount),
                 destination_currency: Some(destination_currency),
@@ -227,6 +236,7 @@ impl From<PayoutsUpdate> for PayoutsUpdateInternal {
                 profile_id,
                 status,
                 confirm,
+                payout_type,
                 ..Default::default()
             },
             PayoutsUpdate::PayoutMethodIdUpdate { payout_method_id } => Self {