+++ crates/router/src/core/payments/flows/session_flow.rs
@@ -1,6 +1,10 @@
 use api_models::payments as payment_types;
 use async_trait::async_trait;
-use common_utils::{ext_traits::ByteSliceExt, request::RequestContent};
+use common_utils::{
+    ext_traits::ByteSliceExt,
+    request::RequestContent,
+    types::{AmountConvertor, StringMajorUnitForConnector},
+};
 use error_stack::{Report, ResultExt};
 use masking::ExposeInterface;
 use router_env::metrics::add_attributes;
@@ -403,15 +407,16 @@ fn get_apple_pay_amount_info(
     label: &str,
     session_data: types::PaymentsSessionData,
 ) -> RouterResult<payment_types::AmountInfo> {
+    let required_amount_type = StringMajorUnitForConnector;
+    let apple_pay_amount = required_amount_type
+        .convert(session_data.minor_amount, session_data.currency)
+        .change_context(errors::ApiErrorResponse::PreconditionFailed {
+            message: "Failed to convert amount to string major unit for applePay".to_string(),
+        })?;
     let amount_info = payment_types::AmountInfo {
         label: label.to_string(),
         total_type: Some("final".to_string()),
-        amount: session_data
-            .currency
-            .to_currency_base_unit(session_data.amount)
-            .change_context(errors::ApiErrorResponse::PreconditionFailed {
-                message: "Failed to convert currency to base unit".to_string(),
-            })?,
+        amount: apple_pay_amount,
     };
 
     Ok(amount_info)
@@ -548,22 +553,21 @@ fn create_gpay_session_token(
                 },
             )
             .collect();
-
+        let required_amount_type = StringMajorUnitForConnector;
+        let google_pay_amount = required_amount_type
+            .convert(
+                router_data.request.minor_amount,
+                router_data.request.currency,
+            )
+            .change_context(errors::ApiErrorResponse::PreconditionFailed {
+                message: "Failed to convert amount to string major unit for googlePay".to_string(),
+            })?;
         let session_data = router_data.request.clone();
         let transaction_info = payment_types::GpayTransactionInfo {
             country_code: session_data.country.unwrap_or_default(),
             currency_code: router_data.request.currency,
             total_price_status: "Final".to_string(),
-            total_price: router_data
-                .request
-                .currency
-                .to_currency_base_unit(router_data.request.amount)
-                .attach_printable(
-                    "Cannot convert given amount to base currency denomination".to_string(),
-                )
-                .change_context(errors::ApiErrorResponse::InvalidDataValue {
-                    field_name: "amount",
-                })?,
+            total_price: google_pay_amount,
         };
 
         let required_shipping_contact_fields =